+++ RELEASES.md
@@ -14,9 +14,9 @@ Version 1.18.0 (unreleased)
     [`Unwrapped`][u-1-18] wrappers.
 
 [bm-c-1]: https://docs.rs/bytemuck/^1/bytemuck/trait.Contiguous.html
-[tf-1-18]: https://docs.rs/fixed/~1.17/fixed/traits/trait.Fixed.html
-[u-1-18]: https://docs.rs/fixed/~1.17/fixed/struct.Unwrapped.html
-[w-1-18]: https://docs.rs/fixed/~1.17/fixed/struct.Wrapping.html
+[tf-1-18]: https://docs.rs/fixed/~1.18/fixed/traits/trait.Fixed.html
+[u-1-18]: https://docs.rs/fixed/~1.18/fixed/struct.Unwrapped.html
+[w-1-18]: https://docs.rs/fixed/~1.18/fixed/struct.Wrapping.html
 
 Version 1.17.0 (2022-08-03)
 ===========================