+++ RELEASES.md
@@ -11,6 +11,8 @@ Version 1.18.0 (unreleased)
   * [`F128Bits`][f128b-1-18] has been deprecated to be replaced by the new
     [`F128`][f128-1-18] which has standard floating-point ordering and various
     classification methods and associated constants.
+  * The [`from_str_dec`][u-fsd-1-18] method was added to the
+    [`Unwrapped`][u-1-18] wrapper.
   * The [`Contiguous`][bm-c-1] trait from the [*bytemuck* crate] was implemented
     for all fixed-point numbers, added as a supertrait to the [`Fixed`][tf-1-18]
     trait, and implemented for the [`Wrapping`][w-1-18] and
@@ -21,6 +23,7 @@ Version 1.18.0 (unreleased)
 [f128b-1-18]: https://docs.rs/fixed/~1.18/fixed/struct.F128Bits.html
 [tf-1-18]: https://docs.rs/fixed/~1.18/fixed/traits/trait.Fixed.html
 [u-1-18]: https://docs.rs/fixed/~1.18/fixed/struct.Unwrapped.html
+[u-fsd-1-128]: https://docs.rs/fixed/~1.18/fixed/struct.Unwrapped.html#method.from_str_dec
 [w-1-18]: https://docs.rs/fixed/~1.18/fixed/struct.Wrapping.html
 
 Version 1.17.0 (2022-08-03)